+++ src/ui/MainWindow.cpp
@@ -50,6 +50,7 @@
 #include "playlist/PlaylistDropImportWorker.h"
 #include "ui/PlaylistImportDialog.h"
 #include "ui/PlaylistView.h"
+#include "ui/ScoringWeightsDialog.h"
 #include "ui/SearchView.h"
 #include "ui/SourceDirectoriesDialog.h"
 #include "ui/SplitterPersistence.h"
@@ -1251,6 +1252,7 @@ MainWindow::MainWindow(AppCore *core, QWidget *parent)
         m_playerBar->setReleaseDeviceVisible(canReleaseOutputDevice());
     });
     connect(m_playerBar, &PlayerBar::linkRootsRequested, this, &MainWindow::configureLinkRoots);
+    connect(m_playerBar, &PlayerBar::scoringWeightsRequested, this, &MainWindow::showScoringWeights);
     connect(m_playerBar, &PlayerBar::genreCurationRequested, this, &MainWindow::showGenreCuration);
     connect(m_playerBar, &PlayerBar::analysisStatusRequested, this, &MainWindow::showAnalysisStatus);
     connect(m_playerBar, &PlayerBar::duplicateCopiesRequested, this, &MainWindow::showDuplicateCopies);
@@ -5429,6 +5431,12 @@ void MainWindow::jumpToTrackInfoAlbum(const QString &artistName, const QString &
 void MainWindow::showGenreCuration()
 {
     GenreCurationDialog dialog(m_database, this);
+    dialog.exec();
+}
+
+void MainWindow::showScoringWeights()
+{
+    ScoringWeightsDialog dialog(m_database, listenHistoryPath(), this);
     dialog.exec();
 }