+++ src/ui/PlayerBar.cpp
@@ -568,6 +568,7 @@ PlayerBar::PlayerBar(QWidget *parent)
     QAction *radioBatchSizeBar = radioMenu->addAction(QStringLiteral("Radio batch size..."));
     connect(radioBatchSizeBar, &QAction::triggered, this, &PlayerBar::radioBatchSizeSettingsRequested);
     QAction *radioShuffleSettings = radioMenu->addAction(QStringLiteral("Radio shuffle percent..."));
+    QAction *scoringWeights = radioMenu->addAction(QStringLiteral("Scoring weights..."));
     radioMenu->addSeparator();
     QAction *genreCuration = radioMenu->addAction(QStringLiteral("Genre curation..."));
     connect(radioMenu, &QMenu::aboutToShow, this, [this]() {
@@ -882,6 +883,7 @@ PlayerBar::PlayerBar(QWidget *parent)
     connect(playbackResume, &QAction::triggered, this, &PlayerBar::playbackResumeRequested);
     connect(libraryShuffleSettings, &QAction::triggered, this, &PlayerBar::libraryShuffleSettingsRequested);
     connect(radioShuffleSettings, &QAction::triggered, this, &PlayerBar::radioShuffleSettingsRequested);
+    connect(scoringWeights, &QAction::triggered, this, &PlayerBar::scoringWeightsRequested);
     connect(genreCuration, &QAction::triggered, this, &PlayerBar::genreCurationRequested);
     connect(rediscoveryMix, &QAction::triggered, this, &PlayerBar::rediscoveryMixRequested);
     connect(deepCutsMix, &QAction::triggered, this, &PlayerBar::deepCutsMixRequested);